Refactor interpolation logic by moving it from markdown.ts to a new file mustache.ts and update imports in promptdom.ts
